Share CryptographicKey base class with all platforms
And divert pseudo-streamining from CryptographicEngine to key classes.
