+++ src/PCLCrypto.Shared.NetFx/PCLCrypto.Shared.NetFx.projitems
@@ -12,7 +12,6 @@
     <Compile Include="$(MSBuildThisFileDirectory)AsymmetricKeyAlgorithmProviderFactory.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)CryptographicBuffer.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)CryptographicEngine.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)CryptographicKey.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)DeriveBytes.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)HashAlgorithmProvider.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)KeyDerivationAlgorithmProvider.cs" />
